Use bfloat16 for Qwen3-TTS to fix CUDA assert error
float16 causes numerical instability during generation.
bfloat16 is the recommended dtype for Qwen3-TTS.

+++ src/apps/qwen3_tts_app.py
@@ -60,7 +60,7 @@ def get_model() -> Qwen3TTSModel:
     global _model
     if _model is None:
         device = "cuda" if torch.cuda.is_available() else "cpu"
-        dtype = torch.float16 if device == "cuda" else torch.float32
+        dtype = torch.bfloat16 if device == "cuda" else torch.float32
         _model = Qwen3TTSModel.from_pretrained(
             QWEN3_HF_MODEL,
             device_map=device,
